requirements.txt
####Cycling init site
django
djangorestframework
gunicorn
django-heroku
djangorestframework_simplejwt
djoser
whitenoise
xlrd
django-ratelimit
django-cors-headers
dj_database_url
Celery

####pywikibot

# This is a PIP 6+ requirements file for using Pywikibot and scripts
#
# All dependencies can be installed using:
# $ sudo pip install -r requirements.txt
#
# It is organised so that simple requirements
# are processed first, and more difficult packages
# are last.
# All dependencies other than requests are optional.
#
# It is good practise to install packages using the system
# package manager if it has a packaged version.  If you are
# unsure, please use pip as described at the top of the file.
#
# To get a list of potential matches, use
#
# $ awk -F '[#>=]' '{print $1}' requirements.txt | xargs yum search
#     or
# $ awk -F '[#>=]' '{print $1}' requirements.txt | xargs apt-cache search

# mandatory; see README-conversion.txt
requests>=2.20.1,<2.22.0; python_version == '3.4'
requests>=2.20.1; python_version != '3.4'
enum34>=1.1.6,!=1.1.8; python_version < '3'
ipaddress ; python_version < '3'
# requests security extra
requests[security] ; python_full_version > '2.7.6' and python_full_version < '2.7.9'

# OAuth support
# mwoauth 0.2.4 is needed because it supports getting identity information
# about the user
mwoauth>=0.2.4,!=0.3.1

# core interwiki_graph.py:
pydot >= 1.2

# wikistats.py and scripts
unicodecsv ; python_version < '3'

# cosmetic_changes and scripts/isbn
python-stdnum >= 1.13

# GUI
Pillow >= 6.2.2, < 7.0.0 ; python_version < '3'
Pillow < 6.0.0 ; python_version == '3.4'
Pillow >= 7.1.0 ; python_version >= '3.5'

# core pagegenerators
google >= 1.7
sseclient >= 0.0.18,!=0.0.23,!=0.0.24

# scripts/flickrripper.py
flickrapi < 3.0.0 ; python_version < '3.5'
flickrapi >= 2.2 ; python_version >= '3.5'

# textlib.py and patrol.py
mwparserfromhell>=0.3.3

# The mysql generator in pagegenerators depends on PyMySQL
PyMySQL

# core HTML comparison parser in diff module
beautifulsoup4

# scripts/states_redirect.py
pycountry

# scripts/weblinkchecker.py
memento_client>=0.5.1,!=0.6.0

# tools/__init__.py
# pywikibot prefers using the inbuilt bz2 module if python was compiled with
# bz2 support. But if it wasn't, bz2file is used instead.
# bz2file

